fix(test): Add bfloat16 dtype to BSHD test model initialization
Fixes dtype mismatch error in tp_only, pp_tp, and all tests where
megatron-bridge's scatter_to_tp_ranks failed due to float32/bfloat16
incompatibility. Other Megatron tests consistently use bfloat16.

+++ tests/models/test_bshd_pp_forward.py
@@ -73,7 +73,7 @@ def create_test_model(tmp_dir: str) -> str:
         num_key_value_heads=2,
         vocab_size=1000,
     )
-    model = AutoModelForCausalLM.from_config(config)
+    model = AutoModelForCausalLM.from_config(config, torch_dtype=torch.bfloat16)
     path = os.path.join(tmp_dir, "test_model")
     model.save_pretrained(path)
     config.save_pretrained(path)
